crud_com(1.0)

+++ sql_app/crud.py
@@ -27,22 +27,34 @@ def get_user(session: Session, user_id: int):
                 }
     return user
 
+
 def add_book(session: Session, user: schems.User, book: schems.Book):
-    db_book = models.Book(title=book.title, author=book.author, user_id=user.id)
+    db_book = models.Book(title=book.title, author=book.author, user_id=user.user_id)
     session.add(db_book)
     return db_book
 
 
-def get_book_by_title(session: Session, title: str):
-    return session.query(models.Book).filter(models.Book.title == title).all()
-
-
-def get_books_by_user(session: Session):
-    return session.query(models.User).get('books')
-
-
 def delete_user(session: Session, user_id: int):
     db_user = session.query(models.User).get(user_id)
+    if db_user == None:
+        return {"user_id": -1,
+                "login": '',
+                "book": []
+                }
     session.delete(db_user)
     session.commit()
     return 'DELETE USER SUCCESSFULLY'
+
+
+def update_user(session: Session, user: schems.UserUpdate):
+    db_user = session.query(models.User).get(user.user_id)
+    if db_user == None:
+        return {"user_id": -1,
+                "login": '',
+                "book": []
+                }
+    db_user.login = user.login
+    db_user.password = user.password
+    session.refresh(db_user)
+    session.commit()
+    return db_user
